--- client/src/app/pages/poam-processing/poam-assigned-grid/poam-assigned-grid.component.html
@@ -1,123 +1,0 @@
-﻿<!--
-!##########################################################################
-! CRANE PLAN OF ACTION AND MILESTONE AUTOMATION TOOL (C-PAT) SOFTWARE
-! Use is governed by the Open Source Academic Research License Agreement
-! contained in the LICENSE.MD file, which is part of this software package.
-! BY USING OR MODIFYING THIS SOFTWARE, YOU ARE AGREEING TO THE TERMS AND
-! CONDITIONS OF THE LICENSE.
-!##########################################################################
--->
-
-<p-table
-  #dt
-  [value]="displayedData()"
-  [columns]="columns"
-  [scrollable]="true"
-  scrollHeight="64rem"
-  sortMode="multiple"
-  class="h-full"
-  selectionMode="single"
-  [globalFilterFields]="['poamId', 'vulnerabilityId', 'affectedAssets', 'scheduledCompletionDate', 'adjSeverity', 'status', 'owner', 'assignedTeams', 'labels', 'associatedVulnerabilities']"
->
-  <ng-template pTemplate="header">
-    <tr>
-      @for (col of columns; track col.field) {
-        @if (col.field === 'status' && col.sortable) {
-          <th [pSortableColumn]="col.field" (click)="onStatusHeaderClick($event)" [pTooltip]="getCurrentStatusSortName()" tooltipPosition="top" scope="col" class="cursor-pointer">
-            {{ col.header }}
-            <i [class]="'mt-1 ml-1 p-sortable-column-icon pi ' + getCurrentStatusSortIcon()" [style.color]="getStatusSortIconColor()" [style.opacity]="0.75"></i>
-          </th>
-        } @else {
-          <th [pSortableColumn]="col.sortable ? col.field : null" [pTooltip]="col.tooltip" tooltipPosition="top" (click)="col.sortable ? resetStatusSort() : null" scope="col">
-            {{ col.header }}
-          </th>
-        }
-      }
-    </tr>
-  </ng-template>
-  <ng-template pTemplate="caption">
-    <div class="flex items-center">
-      <p-iconField iconPosition="left">
-        <p-inputIcon>
-          <i class="pi pi-search"></i>
-        </p-inputIcon>
-        <input id="poamAssignedGridSearch" name="poamAssignedGridSearch" pInputText type="text" [ngModel]="globalFilter" (ngModelChange)="globalFilterSignal.set($event)" placeholder="Search..." />
-      </p-iconField>
-
-      <p-button class="ml-2" icon="pi pi-external-link" severity="secondary" [rounded]="true" [text]="true" (click)="exportToCSV()" pTooltip="Export table data to CSV" tooltipPosition="left" />
-    </div>
-  </ng-template>
-
-  <ng-template pTemplate="body" let-row>
-    <tr>
-      @for (col of columns; track col.field) {
-        <td>
-          @switch (col.field) {
-            @case ('poamId') {
-              <a href="javascript:void(0)" (click)="managePoam(row)" class="text-primary cursor-pointer font-semibold hover:font-bold" pTooltip="Navigate to this POAM" tooltipPosition="top">
-                {{ row.poamId }}
-              </a>
-            }
-
-            @case ('affectedAssets') {
-              @if (row.isAffectedAssetsLoading) {
-                <p-progressSpinner [style]="{ width: '20px', height: '20px' }" strokeWidth="4" [pTooltip]="'Loading affected assets for Vulnerability ID: ' + row.vulnerabilityId" tooltipPosition="top" />
-              } @else if (row.shouldReviewForClosure) {
-                <div class="flex items-center">
-                  <i class="pi pi-exclamation-triangle text-yellow-500 mr-2" [pTooltip]="'Review for closure. No affected assets found for Vulnerability ID: ' + row.vulnerabilityId" tooltipPosition="top"></i>
-                </div>
-              } @else {
-                <div class="flex items-center">
-                  <span>{{ row.affectedAssets }}</span>
-                  @if (row.hasAssociatedVulnerabilities) {
-                    <i class="pi pi-info-circle ml-2 cursor-pointer" style="color: var(--primary-color)" [pTooltip]="row.associatedVulnerabilitiesTooltip" tooltipPosition="top"></i>
-                  }
-                </div>
-              }
-            }
-
-            @case ('assignedTeams') {
-              <div class="flex flex-wrap gap-2 max-h-48 overflow-y-auto overflow-x-hidden">
-                @for (team of row.assignedTeams; track team.name) {
-                  @if (team.name) {
-                    <p-tag [value]="team.name" [rounded]="true" [severity]="getTeamSeverity(team.complete)" [pTooltip]="getTeamTooltip(team.complete)" tooltipPosition="top" />
-                  }
-                }
-              </div>
-            }
-
-            @case ('labels') {
-              <div class="flex flex-wrap gap-2 max-h-48 overflow-y-auto overflow-x-hidden">
-                @for (label of row.labels; track label) {
-                  @if (label) {
-                    <p-tag [value]="label" [rounded]="true" />
-                  }
-                }
-              </div>
-            }
-
-            @case ('associatedVulnerabilities') {
-              <div class="flex flex-wrap gap-2 max-h-48 overflow-y-auto overflow-x-hidden">
-                @for (vuln of row.associatedVulnerabilities; track vuln) {
-                  @if (vuln?.trim()) {
-                    <p-tag [rounded]="true" [value]="vuln.trim()" />
-                  }
-                }
-              </div>
-            }
-
-            @default {
-              {{ row[col.field] }}
-            }
-          }
-        </td>
-      }
-    </tr>
-  </ng-template>
-
-  <ng-template pTemplate="emptymessage">
-    <tr>
-      <td [attr.colspan]="columns.length" class="text-center">No POAMs found.</td>
-    </tr>
-  </ng-template>
-</p-table>
