--- client/src/app/pages/poam-processing/poam-assigned-grid/poam-assigned-grid.component.scss
@@ -1,9 +1,0 @@
-/*
-!##########################################################################
-! CRANE PLAN OF ACTION AND MILESTONE AUTOMATION TOOL (C-PAT) SOFTWARE
-! Use is governed by the Open Source Academic Research License Agreement
-! contained in the LICENSE.MD file, which is part of this software package.
-! BY USING OR MODIFYING THIS SOFTWARE, YOU ARE AGREEING TO THE TERMS AND
-! CONDITIONS OF THE LICENSE.
-!##########################################################################
-*/
