+++ client/src/app/pages/poam-processing/poam-grid/poam-grid.component.scss
@@ -8,9 +8,39 @@
 !##########################################################################
 */
 
+:host {
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+}
+
+::ng-deep .p-card {
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+
+  .p-card-body {
+    display: flex;
+    flex-direction: column;
+    flex: 1;
+    min-height: 0;
+    padding: 0.25rem;
+
+    .p-card-content {
+      display: flex;
+      flex-direction: column;
+      flex: 1;
+      min-height: 0;
+      padding-bottom: 0;
+    }
+  }
+}
+
 .table-container {
-  height: 50rem;
-  overflow-y: auto;
+  display: flex;
+  flex-direction: column;
+  flex: 1;
+  min-height: 0;
   margin-top: 15px;
 }
